Update unit tests and telemetry
